Singletons and Copying
Added base descriptions for Scriptable Object and Monobehaviour Singleton classes. 

Copy Factory provides a method for creating deep copy clones of any value, with the option to customise how some values are treated in the process. 

Added a base class for a Volatile Scriptable Object that makes an (editor only) copy of the objects data so that when play-mode is exited the previous state will be returned. 

Modified the Serializable Interface base to properly handle the displaying of multiple different references and use reflection for type identification instead of real  object value(s).